vim: Fix and move highlight group chain helper
`synIDtrans` follows all the links to the root so we were failing to
capture the ones in the middle.  I still recall this used to work but I
might as well be just imagining.

Unfortunately there's no elegant way to walk the highlight group links
with Vim functions.  Instead chase the links by observing the
':highlight <group name>' commands output and parsing 'xxx links to
<target>'

This does not necessarily belong to vimrc, so I decided to move it to
noclown.vim and define a <Plug> shortcut so it can be bound to and
actual normal mode shortcut from vimrc.