zsh: HISTORY_IGNORE instead of prefixed aliases
Note the change of behavior and policy here:
- Command are still entered history so they can be recalled with history
search while the session is live but will not be persisted to history
file and the end of session.

- Instead of ignoring all `cd` and `ls` commands, we know ignore generic
ones without a parameter.  Granted we set option `hist_ignore_all_dups`,
I'm not sold on the idea how much of clutter we induce anyway but will
keep this in this setting just in case in the future we may want to
expand them with wildcards.