sh: Make shell detection predicates `set -u` safe
When we try to provide Zsh and Bash specific versions in a script, we
rely on checking presence of ZSH_VERSION and BASH_VERSION environment
variables.  If `-u` shell option is set, attempt to expand unset
variables are treated as an error and non-interactive shells will exit.

Use trailing `-` to expand possibly unset variables to empty strings.