bootstrap/termux: support sourcing
Shebang cannot have an environment variable. (Hindsight is 20/20)
Bootstrapper must be run either "bash termux.sh" or "source termux.sh".
Only if it is sourced we can actually switch to zsh by replacing running bash.
Setting '-euo pipefail' while getting sourced will lead to closing Termux
if a command files and we reference an undefined variable. So set it only
when we're executing as a subshell.