vim: 'nolist' for go and man filetypes
Instead of autocmd dance for enabling and disabling tab rendering in
listchars for filetypes where tabs are acceptable, forgo 'list'
entirely. The downside of this is not being able to render trailing
spaces. For Go, this is a non-issue because gofmt on save will take care
of them. No trailing space problem for ft=man as they are unmodifiable
buffers anyway.