.local/bin/proxy: Add output format options
A 'socks5://' url is not enough for Curl as it requires 'socks5h' url
scheme to do the name resolution on the proxy side, nor other apps that
simply expect a host:port pair.

Adds three format options:
  "url" (default) -> 'socks5://127.0.0.1:<listening_port>'
  "curl"          -> '-x socks5h://127.0.0.1:<listening_port>'
  "hostport"      -> '127.0.0.1:<listening_port>'

Also fixes a bug when both tor and TorBrowser were running we wouldn't
default to the first match but output multiple lines from the function,
essentially breaking things.